requirements.txt
flask
discord.py
python-dotenv
requests